Modified the Authentication Controller to have a Logout function
Created a new route called logout in the authentication controller. Also
added functionality to add the logged in user to the currentuser table
and remove the user who is logging off. Added some new fields into the
database table currentusers as well as created a simple procedure in the
DB to log out.